# backend/src/main/resources/openapi/ai-v1.yaml
openapi: 3.1.0
info:
  title: 学生宿舍 AI 控制面 API
  version: 1.0.0
  description: >-
    首期阶段 0-6 的内部管理员 API。所有业务写均通过 proposal、人工审批、
    step-up 和现有 Service；本文档不授权动态 SQL、任意工具或供应商直连。
servers:
  - url: /
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /api/ai/prompts:
    get:
      tags: [Governance]
      operationId: listAiPromptVersions
      parameters:
        - { name: promptKey, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [DRAFT, ACTIVE, INACTIVE, RETIRED] } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [Governance]
      operationId: createAiPromptVersion
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePromptRequest' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/ai/prompts/{id}:
    parameters: [{ $ref: '#/components/parameters/Uuid' }]
    get:
      tags: [Governance]
      operationId: getAiPromptVersion
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '404': { $ref: '#/components/responses/Error' }
  /api/ai/prompts/{id}/activate:
    parameters:
      - $ref: '#/components/parameters/Uuid'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/StepUpProof'
    post:
      tags: [Governance]
      operationId: activateAiPromptVersion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivateSlotRequest' }
      responses:
        '204': { description: active slot 已通过行锁/CAS 原子切换 }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/ai/model-aliases/{alias}/activate:
    parameters:
      - name: alias
        in: path
        required: true
        schema: { type: string, pattern: '^[a-z][a-z0-9._-]{1,63}$' }
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/StepUpProof'
    post:
      tags: [Governance]
      operationId: activateAiModelAlias
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivateModelAliasRequest' }
      responses:
        '204': { description: alias 已按 numeric version CAS 切换 }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /api/ai/tool-catalogs/{id}/activate:
    parameters:
      - $ref: '#/components/parameters/Uuid'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/StepUpProof'
    post:
      tags: [Governance]
      operationId: activateAiToolCatalog
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivateToolCatalogRequest' }
      responses:
        '204': { description: 标准 7 工具目录已原子激活 }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/ai/eval-runs:
    post:
      tags: [Evaluation]
      operationId: createAiEvalRun
      description: >-
        创建时在同一事务预留独立 EVALUATION/offline 预算；后台固定 SERVICE actor 运行
        确定性合同评测。该端点不声称真实供应商或语义质量通过。
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEvalRunRequest' }
      responses:
        '202':
          description: 评测已排队
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Error' }
  /api/ai/eval-runs/{id}:
    parameters: [{ $ref: '#/components/parameters/Uuid' }]
    get:
      tags: [Evaluation]
      operationId: getAiEvalRun
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: Authorization
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Uuid:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
    StepUpProof:
      name: X-Step-Up-Proof
      in: header
      required: true
      schema: { type: string, minLength: 1 }
...
  responses:
...
    Ok:
      description: 成功；响应为 private/no-store
      headers:
        Cache-Control: { schema: { type: string } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiEnvelope' }
    Created:
      description: 已创建
      headers:
        Location: { schema: { type: string } }
        Cache-Control: { schema: { type: string } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiEnvelope' }
...
    Error:
      description: 安全错误合同；不返回堆栈、SQL 或敏感正文
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    RateLimited:
      description: 请求超过 actor/route 策略；errorCode 固定为 AI_RATE_LIMIT_EXCEEDED。
      headers:
        Retry-After:
          required: true
          schema: { type: integer, minimum: 1 }
        X-RateLimit-Limit:
          required: true
          schema: { type: integer, minimum: 1 }
        X-RateLimit-Remaining:
          required: true
          schema: { type: integer, minimum: 0 }
        X-RateLimit-Reset:
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
        X-RateLimit-Policy:
          required: true
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RateLimitErrorEnvelope' }
  schemas:
...
    CreatePromptRequest:
      type: object
      additionalProperties: false
      required: [promptKey, version, content, responseSchemaVersion]
      properties:
        promptKey: { type: string, maxLength: 128 }
        version: { type: string, pattern: '^v[1-9][0-9]{0,8}$' }
        content: { type: string, minLength: 1, maxLength: 100000 }
        responseSchemaVersion: { type: string, maxLength: 64 }
    ActivateSlotRequest:
      type: object
      additionalProperties: false
      properties:
        expectedActiveId: { type: [string, 'null'], format: uuid }
    ActivateModelAliasRequest:
      type: object
      additionalProperties: false
      required: [deploymentId, version]
      properties:
        deploymentId: { type: string, format: uuid }
        version: { type: integer, format: int64, minimum: 0 }
    ActivateToolCatalogRequest:
      type: object
      additionalProperties: false
      required: [version, manifestHash]
      properties:
        version: { type: string, maxLength: 32 }
        manifestHash: { type: string, pattern: '^[0-9a-fA-F]{64}$' }
        expectedActiveId: { type: [string, 'null'], format: uuid }
    CreateEvalRunRequest:
      type: object
      additionalProperties: false
      required: [suiteName, datasetVersion, promptId, modelDeploymentId, codeRevision]
      properties:
        suiteName:
          type: string
          minLength: 3
          maxLength: 128
          description: 服务端只接受已登记的六个离线套件及其精确 datasetVersion 组合。
        datasetVersion: { type: string, maxLength: 64 }
        promptId: { type: string, format: uuid }
        modelDeploymentId: { type: string, format: uuid }
        codeRevision: { type: string, minLength: 3, maxLength: 128 }
    ApiEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code: { type: integer, const: 0 }
        message: { type: string }
        data: {}
    ErrorEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code: { type: integer }
        message: { type: string }
        data:
          type: object
          required: [errorCode, retryable]
          properties:
            errorCode: { type: string }
            retryable: { type: boolean }
            runId: { type: [string, 'null'], format: uuid }
            fieldErrors: { type: array, items: { type: object } }
    RateLimitErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ErrorEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [errorCode, retryable]
              properties:
                errorCode: { type: string, const: AI_RATE_LIMIT_EXCEEDED }
                retryable: { type: boolean, const: true }